Add files via upload
While working for a major real estate and property company, I was entrusted to manage the Air Filters inventory for around 10-12 commercial properties, Each property has multitude of cooling equipment with different sizes for their air filters. Some of the challenges to maintain an inventory were :
1. Every quarterly the filters were ideally replaced but not necessarily replaced due to the pandemic where tenants where working from home.
2. Some filter sizes were common in few property(s), making it challenging to track, order new ones for stock and finally invoicing to the right property
3. Some tenants requested their filters to be upgraded to M13 from the usual M8, due to covid which further made it tricky to keep track of inventory and ordering from supplier.

I created this filter management inventory  on MS Excel using VLOOKUP mostly for the office staff. With this inventory every filter replaced will be deducted from the existing inventory and every filter purchased through a supplier will be added to the stock inventory. And finally when the stock is at 50% of the required yearly numbers, excel will highlight it making it easy for staff to order more filters way ahead of time.